refactor: access store provider from order resource
