+++ hexa_delivery/lib/resources/create_order.dart
@@ -1,9 +1,27 @@
 import 'dart:convert';
 
 import 'package:hexa_delivery/model/dto.dart';
+import 'package:hexa_delivery/resources/store_provider.dart';
 import 'package:http/http.dart' as http;
 
 class OrderResource {
+  Map<String, StoreDTO>? _storeNameDTOMap;
+
+  bool isCreated(String query) {
+    if (_storeNameDTOMap == null) return true;
+    // The query is guaranteed to there (search before suggest & select)
+    if (_storeNameDTOMap![query]!.isFromAPI()) return false;
+    return true;
+  }
+
+  Future<Iterable<String>> getStoreNameList(String query) async {
+    List<StoreDTO> storeList = await StoreListQueryProvider.searchStoresAndGetList(query);
+    Iterable<MapEntry<String, StoreDTO>> entries = storeList.map((store) => MapEntry(store.getName, store));
+    _storeNameDTOMap = Map.fromEntries(entries); // cache store map
+
+    return storeList.map((store) => store.getName);
+  }
+
   void createOrder(OrderToBeCreatedDTO order, User user) async {
     var headers = {
       "Access-Token": user.getToken(),
